Feature: apollo_router::main
This pull request exposes apollo_router main function and it's async variant so plugin writers can spin up a fully fledged router with their plugin in 5 lines of rust.
