+++ apollo-router/src/lib.rs
@@ -2,6 +2,7 @@
 
 mod apollo_telemetry;
 pub mod configuration;
+mod executable;
 mod files;
 mod http_server_factory;
 mod layers;
@@ -20,6 +21,7 @@ use configuration::{Configuration, ListenAddr};
 use derivative::Derivative;
 use derive_more::{Display, From};
 use displaydoc::Display as DisplayDoc;
+pub use executable::main;
 use futures::channel::{mpsc, oneshot};
 use futures::prelude::*;
 use futures::FutureExt;
