+++ examples/async-allow-client-id/Cargo.toml
@@ -11,7 +11,6 @@ apollo-router = { path = "../../apollo-router" }
 apollo-router-core = { path = "../../apollo-router-core" }
 futures = "0.3.21"
 http = "0.2.6"
-pin-project = "1.0.10"
 schemars = "0.8.8"
 serde = "1.0.136"
 serde_json = "1.0.79"
