+++ examples/forbid-anonymous-operations/config.yml
